Language.Java.Java

Documentation

module Language.Java.Syntax

module Language.Java.SyntaxClasses

data Type

There are two kinds of types in the Java programming language: primitive types and reference types.

Constructors

PrimType PrimType

RefType RefType

data RefType

There are three kinds of reference types: class types, interface types, and array types. Reference types may be parameterized with type arguments. Type variables cannot be syntactically distinguished from class type identifiers, and are thus represented uniformly as single ident class types.

Constructors

ClassRefType ClassType

ArrayType Type
TypeVariable Ident

data ClassType

A ClassType can either be with an package or without.

Constructors

WithPackage Package ClassName

WithoutPackage ClassName

data ClassName

a Classname is either a Classname or wildcard (matches any name). An AST constructed by the parser will only produce Classnames.

Constructors

ClassName [(Ident, [TypeArgument])]

WildcardName

data Package

represents a package, e.g. java.util a package can either be fully qualified ("java.util"), or end with an wildcard ("java.util.*")

Constructors

FullQualiPackage [Ident]

WildcardPackage [Ident]

data TypeArgument

Type arguments may be either reference types or wildcards.

Constructors

Wildcard (Maybe WildcardBound)

ActualType RefType

Diamond

newtype TypeDeclSpecifier

Constructors

TypeDeclSpecifier ClassType

data WildcardBound

Wildcards may be given explicit bounds, either upper (extends) or lower (super) bounds.

Constructors

ExtendsBound RefType

SuperBound RefType

data PrimType

A primitive type is predefined by the Java programming language and named by its reserved keyword.

Constructors

BooleanT

ByteT

ShortT

IntT

LongT

CharT

FloatT

DoubleT

data TypeParam

A class is generic if it declares one or more type variables. These type variables are known as the type parameters of the class.

Constructors

TypeParam

Fields

typeParamName :: Ident

typeParamBounds :: [RefType]

newtype Ident

A single identifier.

Constructors

Ident String

fromIdent :: Ident -> String

newtype Name

A name, i.e. a period-separated list of identifiers.

Constructors

Name [Ident]

newtype RelaxedType

A type with relaxed equality checking: 1. If packages equal 2. Or if classes equal 3. Or if boxed primitives equal primitives

Constructors

RelaxedType Type

checkRelaxed :: RefType -> RefType -> Bool

Checks a RelaxedType for equality. A RelaxedType equals another RelaxedType if: 1. If packages equal 2. If classes equal 3. If boxed primitives equal primitives

primToRefType :: PrimType -> RefType

This function returns a primitve as a ref type (i.e. boxed primitve)

withPackageIdentToType :: [Ident] -> Ident -> Type

This function returns an Ident as a Type

withoutPackageIdentToType :: Ident -> Type

This function returns an Ident as a Type

data Literal

A literal denotes a fixed, unchanging value.

Constructors

Int Integer

Word Integer

Float Double

Double Double

Boolean Bool

Char Char

String String

Null

data Op

A binary infix operator.

Constructors

Mult

Div

Rem

Add

Sub

LShift

RShift

RRShift

LThan

GThan

LThanE

GThanE

Equal

NotEq

And

Or

Xor

CAnd

COr

data AssignOp

An assignment operator.

Constructors

EqualA

MultA

DivA

RemA

AddA

SubA

LShiftA

RShiftA

RRShiftA

AndA

XorA

OrA