Language.Java.Syntax

Documentation

data CompilationUnitNode l

A compilation unit is the top level syntactic goal symbol of a Java program.

Constructors

CompilationUnitNode (CompilationUnit l)

ModuleDeclarationNode (ModuleDeclaration l)

data ModuleSpecNode l

specifies the module declarations

Constructors

ModuleRequiresNode (ModuleRequires l)
requires the module to work

ModuleExportsNode (ModuleExports l)
exports the package

data TypeDeclNode l

A type declaration declares a class type or an interface type.

Constructors

ClassTypeDeclNode (ClassDeclNode l)

InterfaceTypeDeclNode (InterfaceDecl l)

data ClassDeclNode l

A class declaration specifies a new named reference type.

Constructors

ClassDeclNode (ClassDecl l)

EnumDeclNode (EnumDecl l)

data DeclNode l

A declaration is either a member declaration, or a declaration of an initializer, which may be static.

Constructors

MemberDeclNode (MemberDeclNode l)

InitDeclNode (InitDecl l)

data MemberDeclNode l

A class or interface member can be an inner class or interface, a field or constant, or a method or constructor. An interface may only have as members constants (not fields), abstract methods, and no constructors.

Constructors

FieldDeclNode (FieldDecl l)
The variables of a class type are introduced by field declarations.

MethodDeclNode (MethodDecl l)
A method declares executable code that can be invoked, passing a fixed number of values as arguments.

ConstructorDeclNode (ConstructorDecl l)
A constructor is used in the creation of an object that is an instance of a class.

MemberClassDeclNode (ClassDeclNode l)
A member class is a class whose declaration is directly enclosed in another class or interface declaration.

MemberInterfaceDeclNode (InterfaceDecl l)
A member interface is an interface whose declaration is directly enclosed in another class or interface declaration.

data VarDeclIdNode l

The name of a variable in a declaration, which may be an array.

Constructors

VarIdNode (VarId l)

VarDeclArrayNode (VarDeclIdNode l)
Multi-dimensional arrays are represented by nested applications of VarDeclArray.

data VarInitNode l

Explicit initializer for a variable declaration.

Constructors

InitExpNode (ExpNode l)

InitArrayNode (ArrayInit l)

data ExplConstrInvNode l

An explicit constructor invocation invokes another constructor of the same class, or a constructor of the direct superclass, which may be qualified to explicitly specify the newly created object's immediately enclosing instance.

Constructors

ThisInvokeNode (ThisInvoke l)

SuperInvokeNode (SuperInvoke l)

PrimarySuperInvokeNode (PrimarySuperInvoke l)

data BlockStmtNode l

A block statement is either a normal statement, a local class declaration or a local variable declaration.

Constructors

BlockStmtNode (StmtNode l)

LocalClassNode (ClassDeclNode l)

LocalVarsNode (LocalVars l)

data StmtNode l

A Java statement.

Constructors

StmtBlockNode (Block l)
A statement can be a nested block.

IfThenElseNode (IfThenElse l)
The if-then statement allows conditional execution of a statement.

WhileNode (While l)
The while statement executes an expression and a statement repeatedly until the value of the expression is false.

BasicForNode (BasicFor l)
The basic for statement executes some initialization code, then executes an expression, a statement, and some update code repeatedly until the value of the expression is false.

EnhancedForNode (EnhancedFor l)
The enhanced for statement iterates over an array or a value of a class that implements the iterator interface.

EmptyNode (Empty l)
An empty statement does nothing.

ExpStmtNode (ExpNode l)
Certain kinds of expressions may be used as statements by following them with semicolons: assignments, pre- or post-inc- or decrementation, method invocation or class instance creation expressions.

AssertNode (Assert l)
An assertion is a statement containing a boolean expression, where an error is reported if the expression evaluates to false.

SwitchNode (Switch l)
The switch statement transfers control to one of several statements depending on the value of an expression.

DoNode (Do l)
The do statement executes a statement and an expression repeatedly until the value of the expression is false.

BreakNode (Break l)
A break statement transfers control out of an enclosing statement.

ContinueNode (Continue l)
A continue statement may occur only in a while, do, or for statement. Control passes to the loop-continuation point of that statement.

ReturnNode (Return l)

SynchronizedNode (Synchronized l)
A synchronized statement acquires a mutual-exclusion lock on behalf of the executing thread, executes a block, then releases the lock. While the executing thread owns the lock, no other thread may acquire the lock.

ThrowNode (Throw l)
A throw statement causes an exception to be thrown.

TryNode (Try l)
A try statement executes a block and may catch a thrown exception

LabeledNode (Labeled l)
Statements may have label prefixes.

data TryResourceNode l

Resource in a try-with-resources statement

Constructors

TryResourceVarNode (TryResourceVar l)
Newly declared variables

TryResourceFinalVarNode (TryResourceFinalVar l)
Effectively final variable

data SwitchLabelNode l

A label within a switch statement.

Constructors

SwitchCaseNode (ExpNode l)
The expression contained in the case must be a Lit or an enum constant.

DefaultNode l

data ForInitNode l

Initialization code for a basic for statement.

Constructors

ForLocalVarsNode (ForLocalVars l)

ForInitExpsNode (ForInitExps l)

data ExpNode l

A Java expression.

Constructors

LitNode (Lit l)
A literal denotes a fixed, unchanging value.

ClassLitNode (ClassLit l)
A class literal, which is an expression consisting of the name of a class, interface, array, or primitive type, or the pseudo-type void (modelled by Nothing), followed by a . and the token class.

ThisNode (This l)
The keyword this denotes a value that is a reference to the object for which the instance method was invoked, or to the object being constructed.

QualifiedThisNode (QualifiedThis l)
Any lexically enclosing instance can be referred to by explicitly qualifying the keyword this. TODO: Fix Parser here

InstanceCreationNode (InstanceCreation l)
A class instance creation expression is used to create new objects that are instances of classes. | The first argument is a list of non-wildcard type arguments to a generic constructor. What follows is the type to be instantiated, the list of arguments passed to the constructor, and optionally a class body that makes the constructor result in an object of an anonymous class.

QualInstanceCreationNode (QualInstanceCreation l)
A qualified class instance creation expression enables the creation of instances of inner member classes and their anonymous subclasses.

ArrayCreateNode (ArrayCreate l)
An array instance creation expression is used to create new arrays. The last argument denotes the number of dimensions that have no explicit length given. These dimensions must be given last.

ArrayCreateInitNode (ArrayCreateInit l)
An array instance creation expression may come with an explicit initializer. Such expressions may not be given explicit lengths for any of its dimensions.

FieldAccessNode (FieldAccessNode l)
A field access expression.

MethodInvNode (MethodInvocationNode l)
A method invocation expression.

ArrayAccessNode (ArrayIndex l)
An array access expression refers to a variable that is a component of an array.

ExpNameNode (ExpName l)
An expression name, e.g. a variable.

PostIncrementNode (ExpNode l)
Post-incrementation expression, i.e. an expression followed by ++.

PostDecrementNode (ExpNode l)
Post-decrementation expression, i.e. an expression followed by --.

PreIncrementNode (ExpNode l)
Pre-incrementation expression, i.e. an expression preceded by ++.

PreDecrementNode (ExpNode l)
Pre-decrementation expression, i.e. an expression preceded by --.

PrePlusNode (ExpNode l)
Unary plus, the promotion of the value of the expression to a primitive numeric type.

PreMinusNode (ExpNode l)
Unary minus, the promotion of the negation of the value of the expression to a primitive numeric type.

PreBitComplNode (ExpNode l)
Unary bitwise complementation: note that, in all cases, ~x equals (-x)-1.

PreNotNode (ExpNode l)
Logical complementation of boolean values.

CastNode (Cast l)
A cast expression converts, at run time, a value of one numeric type to a similar value of another numeric type; or confirms, at compile time, that the type of an expression is boolean; or checks, at run time, that a reference value refers to an object whose class is compatible with a specified reference type.

BinOpNode (BinOp l)
The application of a binary operator to two operand expressions.

InstanceOfNode (InstanceOf l)
Testing whether the result of an expression is an instance of some reference type.

CondNode (Cond l)
The conditional operator ? : uses the boolean value of one expression to decide which of two other expressions should be evaluated.

AssignNode (Assign l)
Assignment of the result of an expression to a variable.

LambdaNode (Lambda l)
Lambda expression

MethodRefNode (MethodRef l)
Method reference

data LhsNode l

The left-hand side of an assignment expression. This operand may be a named variable, such as a local variable or a field of the current object or class, or it may be a computed variable, as can result from a field access or an array access.

Constructors

NameLhsNode (NameLhs l)
Assign to a variable

FieldLhsNode (FieldAccessNode l)
Assign through a field access

ArrayLhsNode (ArrayIndex l)
Assign to an array

data FieldAccessNode l

A field access expression may access a field of an object or array, a reference to which is the value of either an expression or the special keyword super.

Constructors

PrimaryFieldAccessNode (PrimaryFieldAccess l)
Accessing a field of an object or array computed from an expression.

SuperFieldAccessNode (SuperFieldAccess l)
Accessing a field of the superclass.

ClassFieldAccessNode (ClassFieldAccess l)
Accessing a (static) field of a named class.

data LambdaParamsNode l

Constructors

LambdaSingleParamNode (LambdaSingleParam l)

LambdaFormalParamsNode (LambdaFormalParams l)

LambdaInferredParamsNode (LambdaInferredParams l)

data LambdaExpressionNode l

Lambda expression, starting from java 8

Constructors

LambdaExpressionNode (ExpNode l)

LambdaBlockNode (Block l)

data MethodInvocationNode l

A method invocation expression is used to invoke a class or instance method.

Constructors

MethodCallNode (MethodCall l)
Invoking a specific named method.

PrimaryMethodCallNode (PrimaryMethodCall l)
Invoking a method of a class computed from a primary expression, giving arguments for any generic type parameters.

SuperMethodCallNode (SuperMethodCall l)
Invoking a method of the super class, giving arguments for any generic type parameters.

ClassMethodCallNode (ClassMethodCall l)
Invoking a method of the superclass of a named class, giving arguments for any generic type parameters.

TypeMethodCallNode (TypeMethodCall l)
Invoking a method of a named type, giving arguments for any generic type parameters.

data CompilationUnit l

A compilation unit is the top level syntactic goal symbol of a Java program.

Constructors

CompilationUnit

Fields

infoCompUnit :: l

packageLocation :: Maybe (PackageDecl l)

imports :: [ImportDecl l]

typeDecls :: [TypeDeclNode l]

data ModuleDeclaration l

Constructors

ModuleDeclaration

Fields

infoModuleDecl :: l

modulePackage :: Package

moduleSpecs :: [ModuleSpecNode l]

data PackageDecl l

A package declaration appears within a compilation unit to indicate the package to which the compilation unit belongs.

Constructors

PackageDecl

Fields

infoPackDec :: l

packageDecl :: Package

data ModuleRequires l

requires the module to work

Constructors

ModuleRequires

Fields

infoModuleRequires :: l

requireModule :: Package

data ModuleExports l

Exports the package

Constructors

ModuleExports

Fields

infoModuleExports :: l

exportsPackage :: Package

data ImportDecl l

An import declaration allows a static member or a named type to be referred to by a single unqualified identifier. The first argument signals whether the declaration only imports static members. The last argument signals whether the declaration brings all names in the named type or package, or only brings a single name into scope.

Constructors

ImportDecl

Fields

infoImportDecl :: l

staticImport :: Bool

importPackage :: Package

data ClassDecl l

A class declaration specifies a new named reference type.

Constructors

ClassDecl

Fields

infoClassDecl :: l

classDeclModifiers :: [Modifier l]

classDeclName :: Ident

classTypeParams :: [TypeParam]

extends :: Maybe (Extends l)

classImplements :: [Implements l]

classBody :: ClassBody l

data EnumDecl l

Constructors

EnumDecl

Fields

infoEnumDecl :: l

enumDeclModifiers :: [Modifier l]

enumeDeclName :: Ident

enumImplements :: [Implements l]

enumBody :: EnumBody l

data Extends l

An extends clause

Constructors

Extends

Fields

infoExtends :: l

extendsClass :: RefType

data Implements l

An implements clause

Constructors

Implements

Fields

infoImplements :: l

implementsInterface :: RefType

data ClassBody l

A class body may contain declarations of members of the class, that is, fields, classes, interfaces and methods. A class body may also contain instance initializers, static initializers, and declarations of constructors for the class.

Constructors

ClassBody

Fields

infoClassBody :: l

classDecls :: [DeclNode l]

data EnumBody l

The body of an enum type may contain enum constants.

Constructors

EnumBody

Fields

infoEnumBody :: l

enumConstants :: [EnumConstant l]

enumDecls :: [DeclNode l]

data EnumConstant l

An enum constant defines an instance of the enum type.

Constructors

EnumConstant

Fields

infoEnumConstant :: l

enumConstantName :: Ident

enumArguments :: [Argument l]

enumConstantBody :: Maybe (ClassBody l)

data InterfaceDecl l

An interface declaration introduces a new reference type whose members are classes, interfaces, constants and abstract methods. This type has no implementation, but otherwise unrelated classes can implement it by providing implementations for its abstract methods.

Constructors

InterfaceDecl

Fields

infoInterfaceDecl :: l

interfaceKind :: InterfaceKind

interfaceDeclModifiers :: [Modifier l]

interfaceDeclName :: Ident

interfaceTypeParams :: [TypeParam]

interfaceExtends :: [Extends l]

interfaceBody :: InterfaceBody l

data InterfaceKind

Interface can declare either a normal interface or an annotation

Constructors

InterfaceNormal

InterfaceAnnotation

data InterfaceBody l

The body of an interface may declare members of the interface.

Constructors

InterfaceBody

Fields

infoInterfaceBody :: l

members :: [MemberDeclNode l]

data InitDecl l

A declaration is either a member declaration, or a declaration of an initializer, which may be static.

Constructors

InitDecl

Fields

infoInitDecl :: l

staticDecl :: Bool

statements :: Block l

data FieldDecl l

The variables of a class type are introduced by field declarations.

Example:

>>> parseCompilationUnit "public class MyClass {private String foo = \"Hello World\"; }"
...
__FieldDecl__ {__infoFieldDecl__ = Segment (Position 1 31) (Position 1 31), __memberDeclModifiers__ = [private], __fieldType__ =
RefType (ClassRefType (WithoutPackage (ClassName [(Ident "String",[])]))), __fieldVarDecls__ = [VarDecl {infoVarDecl =
Segment (Position 1 38) (Position 1 38), varDeclName = VarId {infoVarId = Segment (Position 1 38) (Position 1 38),
varIdName = Ident "foo"}, varInit = Just (InitExp {infoInitExp= Segment (Position 1 44) (Position 1 44), init = Lit
{infoLit = Segment (Position 1 44) (Position 1 44), literal = String "Hello World"}})}]}}]}}}]})

Constructors

FieldDecl

Fields

infoFieldDecl :: l

memberDeclModifiers :: [Modifier l]

fieldType :: Type

fieldVarDecls :: [VarDecl l]

data MethodDecl l

A method declares executable code that can be invoked, passing a fixed number of values as arguments. Example:

>>> parseCompilationUnit "public class MyClass {private String foo() {}}"
...
[MemberDecl {infoMemberDecl = Segment (Position 1 23) (Position 1 23), member = MethodDecl {infoMethodDecl =
Segment (Position 1 31) (Position 1 31), methodDeclModifiers = [private], methodTypeParams = [], returnType =
Just (RefType (ClassRefType (WithoutPackage (ClassName [(Ident "String",[])])))), methodDeclName = Ident "foo", params = [],
exceptions = [], defaultInterfaceAnnotation = Nothing, methodBody = MethodBody {infoMethodBody = Segment (Position 1 44)
(Position 1 44), impl= Just (Block {infoBlock = Segment (Position 1 45) (Position 1 45), blockStatements = []})}}}]}}}]})

Constructors

MethodDecl

Fields

infoMethodDecl :: l

methodDeclModifiers :: [Modifier l]

methodTypeParams :: [TypeParam]

returnType :: Maybe Type

methodDeclName :: Ident

params :: [FormalParam l]

exceptions :: [ExceptionType l]

defaultInterfaceAnnotation :: Maybe (ExpNode l)

methodBody :: MethodBody l

data ConstructorDecl l

A constructor is used in the creation of an object that is an instance of a class.

Constructors

ConstructorDecl

Fields

infoConstructorDecl :: l

constructorMod :: [Modifier l]

constructorTypeParams :: [TypeParam]

constructorClassName :: Ident

constructorFormalParams :: [FormalParam l]

constructorExceptions :: [ExceptionType l]

constructorBody :: ConstructorBody l

data VarDecl l

A declaration of a variable, which may be explicitly initialized.

Constructors

VarDecl

Fields

infoVarDecl :: l

varDeclName :: VarDeclIdNode l

varInit :: Maybe (VarInitNode l)

data VarId l

The name of a variable in a declaration, which may be an array.

Constructors

VarId

Fields

infoVarId :: l

varIdName :: Ident

data FormalParam l

A formal parameter in method declaration. The last parameter for a given declaration may be marked as variable arity, indicated by the boolean argument.

Constructors

FormalParam

Fields

infoFormalParam :: l

formalParamModifiers :: [Modifier l]

paramType :: Type

variableArity :: Bool

paramName :: VarDeclIdNode l

data MethodBody l

A method body is either a block of code that implements the method or simply a semicolon, indicating the lack of an implementation (modelled by Nothing).

Constructors

MethodBody

Fields

infoMethodBody :: l

impl :: Maybe (Block l)

data ConstructorBody l

The first statement of a constructor body may be an explicit invocation of another constructor of the same class or of the direct superclass.

Constructors

ConstructorBody

Fields

infoConstructorBody :: l

constructorInvoc :: Maybe (ExplConstrInvNode l)

constrBody :: [BlockStmtNode l]

data ThisInvoke l

An explicit constructor invocation invokes another constructor of the same class, or a constructor of the direct superclass, which may be qualified to explicitly specify the newly created object's immediately enclosing instance.

Constructors

ThisInvoke

Fields

infoThisInvoke :: l

thisTypeArguments :: [RefType]

thisConstrArguments :: [Argument l]

data SuperInvoke l

Constructors

SuperInvoke

Fields

infoSuperInvoke :: l

superTypeArguments :: [RefType]

superConstrArguments :: [Argument l]

data PrimarySuperInvoke l

Constructors

PrimarySuperInvoke

Fields

infoPrimarySuperInvoke :: l

primary :: ExpNode l

primaryTypeArguments :: [RefType]

primaryConstrArguments :: [Argument l]

data Modifier l

A modifier specifying properties of a given declaration. In general only a few of these modifiers are allowed for each declaration type, for instance a member type declaration may only specify one of public, private or protected.

Constructors

Public l

Private l

Protected l

Abstract l

Final l

Static l

StrictFP l

Transient l

Volatile l

Native l

Annotation (Annotation l)

Synchronized_ l

DefaultModifier l

data Annotation l

Annotations have three different forms: no-parameter, single-parameter or key-value pairs

Constructors

NormalAnnotation

Fields

annName :: Name

annKV :: [(Ident, ElementValue l)]

SingleElementAnnotation

Fields

annName :: Name

annValue :: ElementValue l

MarkerAnnotation

Fields

annName :: Name

data ElementValue l

Annotations may contain annotations or (loosely) expressions

Constructors

EVVal

Fields

infoEVVal :: l

elementVarInit :: VarInitNode l

EVAnn

Fields

infoEVAnn :: l

annotation :: Annotation l

data Block l

A block is a sequence of statements, local class declarations and local variable declaration statements within braces.

Constructors

Block

Fields

infoBlock :: l

blockStatements :: [BlockStmtNode l]

data LocalVars l

A block statement is either a normal statement, a local class declaration or a local variable declaration.

Constructors

LocalVars

Fields

infoLocalVars :: l

locaVarModifiers :: [Modifier l]

blockVarType :: Type

localVarDecls :: [VarDecl l]

data IfThenElse l

The if-then statement allows conditional execution of a statement.

Constructors

IfThenElse

Fields

infoIfThenElse :: l

ifExp :: ExpNode l

thenExp :: StmtNode l

elseExp :: Maybe (StmtNode l)

data While l

The while statement executes an expression and a statement repeatedly until the value of the expression is false.

Constructors

While

Fields

infoWhile :: l

whileVondition :: ExpNode l

whileBody :: StmtNode l

data BasicFor l

The basic for statement executes some initialization code, then executes an expression, a statement, and some update code repeatedly until the value of the expression is false.

Constructors

BasicFor

Fields

infoBasicFor :: l

forInit :: Maybe (ForInitNode l)

forCond :: Maybe (ExpNode l)

forUpdate :: Maybe [ExpNode l]

basicForBody :: StmtNode l

data EnhancedFor l

The enhanced for statement iterates over an array or a value of a class that implements the iterator interface.

Constructors

EnhancedFor

Fields

infoEnhancedFor :: l

loopVarModifiers :: [Modifier l]

example: for (final Int x : set) {..}

loopVarType :: Type

loopVarName :: Ident

iterable :: ExpNode l

enhancedForBody :: StmtNode l

newtype Empty l

An empty statement does nothing.

Constructors

Empty

Fields

infoEmpty :: l

data Assert l

An assertion is a statement containing a boolean expression, where an error is reported if the expression evaluates to false.

Constructors

Assert

Fields

infoAssert :: l

booleanExp :: ExpNode l

valueExp :: Maybe (ExpNode l)

data Switch l

The switch statement transfers control to one of several statements depending on the value of an expression.

Constructors

Switch

Fields

infoSwitch :: l

switchValue :: ExpNode l

switchBlocks :: [SwitchBlock l]

data Do l

The do statement executes a statement and an expression repeatedly until the value of the expression is false.

Constructors

Do

Fields

infoDo :: l

doBody :: StmtNode l

doCondition :: ExpNode l

data Break l

A break statement transfers control out of an enclosing statement.

Constructors

Break

Fields

infoBreak :: l

breakLabel :: Maybe Ident

data Continue l

A continue statement may occur only in a while, do, or for statement. Control passes to the loop-continuation point of that statement.

Constructors

Continue

Fields

infoContinue :: l

continueLabel :: Maybe Ident

data Return l

A return statement returns control to the invoker of a method or constructor.

Constructors

Return

Fields

infoReturn :: l

returnExp :: Maybe (ExpNode l)

data Synchronized l

A synchronized statement acquires a mutual-exclusion lock on behalf of the executing thread, executes a block, then releases the lock. While the executing thread owns the lock, no other thread may acquire the lock.

Constructors

Synchronized

Fields

infoSynchronized :: l

synchronizeOn :: ExpNode l

synchronizeBloc :: Block l

data Throw l

A throw statement causes an exception to be thrown.

Constructors

Throw

Fields

infoThrow :: l

throwExp :: ExpNode l

data Try l

A try statement executes a block. If a value is thrown and the try statement has one or more catch clauses that can catch it, then control will be transferred to the first such catch clause. If the try statement has a finally clause, then another block of code is executed, no matter whether the try block completes normally or abruptly, and no matter whether a catch clause is first given control.

Constructors

Try

Fields

infoTry :: l

tryResource :: [TryResourceNode l]

tryBlock :: Block l

catches :: [Catch l]

finally :: Maybe (Block l)

data Labeled l

Statements may have label prefixes.

Constructors

Labeled

Fields

infoLabeled :: l

label :: Ident

labeledStmt :: StmtNode l

data Catch l

If a value is thrown and the try statement has one or more catch clauses that can catch it, then control will be transferred to the first such catch clause.

Constructors

Catch

Fields

infoCatch :: l

catchParam :: FormalParam l

catchBlock :: Block l

data TryResourceVar l

Newly declared variables

Constructors

TryResourceVar

Fields

infoTryResourceVar :: l

resourceModifiers :: [Modifier l]

resourceVarType :: RefType

resourceVarDecl :: [VarDecl l]

data TryResourceFinalVar l

Effectively final variable

Constructors

TryResourceFinalVar

Fields

infoTryResourceFinalVar :: l

resourceFinalVarName :: Ident

data SwitchBlock l

A block of code labelled with a case or default within a switch statement.

Constructors

SwitchBlock

Fields

infoSwitchBlock :: l

switchLabel :: SwitchLabelNode l

switchStmts :: [BlockStmtNode l]

data ForLocalVars l

Initialization code for a basic for statement.

Constructors

ForLocalVars

Fields

infoForLocalVars :: l

forVarModifiers :: [Modifier l]

forVarType :: Type

forVarDecls :: [VarDecl l]

data ForInitExps l

Constructors

ForInitExps

Fields

infoForInitExps :: l

initExpr :: [ExpNode l]

data ExceptionType l

An exception type has to be a class type or a type variable.

Constructors

ExceptionType

Fields

infoExceptionType :: l

expectionType :: RefType

data Lit l

A literal denotes a fixed, unchanging value.

Constructors

Lit

Fields

infoLit :: l

literal :: Literal

data ClassLit l

A class literal, which is an expression consisting of the name of a class, interface, array, or primitive type, or the pseudo-type void (modelled by Nothing), followed by a . and the token class.

Constructors

ClassLit

Fields

infoClassLit :: l

classLit :: Maybe Type

newtype This l

The keyword this denotes a value that is a reference to the object for which the instance method was invoked, or to the object being constructed.

Constructors

This

Fields

infoThis :: l

data QualifiedThis l

Any lexically enclosing instance can be referred to by explicitly qualifying the keyword this. TODO: Fix Parser here

Constructors

QualifiedThis

Fields

infoQualifiedThis :: l

qualiType :: Type

data InstanceCreation l

A class instance creation expression is used to create new objects that are instances of classes. | The first argument is a list of non-wildcard type arguments to a generic constructor. What follows is the type to be instantiated, the list of arguments passed to the constructor, and optionally a class body that makes the constructor result in an object of an anonymous class.

Constructors

InstanceCreation

Fields

infoInstanceCreation :: l

instanceTypeArgs :: [TypeArgument]

instanceTypeDecl :: TypeDeclSpecifier

instanceArguments :: [Argument l]

anonymousClass :: Maybe (ClassBody l)

data QualInstanceCreation l

A qualified class instance creation expression enables the creation of instances of inner member classes and their anonymous subclasses.

Constructors

QualInstanceCreation

Fields

infoQualInstanceCreation :: l

mysteryExp :: ExpNode l

qualiInstancetypeArgs :: [TypeArgument]

className :: Ident

qualiInstanceArguments :: [Argument l]

qualiInstanceAnonymousClass :: Maybe (ClassBody l)

data ArrayCreate l

An array instance creation expression is used to create new arrays. The last argument denotes the number of dimensions that have no explicit length given. These dimensions must be given last.

Constructors

ArrayCreate

Fields

infoArrayCreate :: l

arrayType :: Type

arrayDimExprs :: [ExpNode l]

dimensions :: Int

data ArrayCreateInit l

An array instance creation expression may come with an explicit initializer. Such expressions may not be given explicit lengths for any of its dimensions.

Constructors

ArrayCreateInit

Fields

infoArrayCreateInit :: l

arrayInitType :: Type

arrayInitDimensions :: Int

arrayCreatInit :: ArrayInit l

data ExpName l

An expression name, e.g. a variable.

Constructors

ExpName

Fields

infoExpName :: l

expName :: Name

data Cast l

A cast expression converts, at run time, a value of one numeric type to a similar value of another numeric type; or confirms, at compile time, that the type of an expression is boolean; or checks, at run time, that a reference value refers to an object whose class is compatible with a specified reference type.

Constructors

Cast

Fields

infoCast :: l

castTarget :: Type

castArg :: ExpNode l

data BinOp l

The application of a binary operator to two operand expressions.

Constructors

BinOp

Fields

infoBinOp :: l

binArgLeft :: ExpNode l

binOp :: Op

binOpRight :: ExpNode l

data InstanceOf l

Testing whether the result of an expression is an instance of some reference type.

Constructors

InstanceOf

Fields

infoInstanceOf :: l

instanceOfArg :: ExpNode l

instanceOfTarget :: RefType

data Cond l

The conditional operator ? : uses the boolean value of one expression to decide which of two other expressions should be evaluated.

Constructors

Cond

Fields

infoCond :: l

condition :: ExpNode l

conditionTrueExp :: ExpNode l

conditionFalseExp :: ExpNode l

data Assign l

Assignment of the result of an expression to a variable.

Constructors

Assign

Fields

infoAssign :: l

assignTarget :: LhsNode l

assignOp :: AssignOp

assignSource :: ExpNode l

data Lambda l

Lambda expression

Constructors

Lambda

Fields

infoLambda :: l

lambdaParams :: LambdaParamsNode l

lambdaExpression :: LambdaExpressionNode l

data MethodRef l

Method reference

Constructors

MethodRef

Fields

infoMethodRef :: l

methodClass :: Name

methodName :: Ident

data NameLhs l

Assign to a variable

Constructors

NameLhs

Fields

infoNameLhs :: l

varLhsName :: Name

data ArrayIndex l

Array access

Constructors

ArrayIndex

Fields

infoArrayIndex :: l

arrayName :: ExpNode l

arrayIndices :: [ExpNode l]

data PrimaryFieldAccess l

Accessing a field of an object or array computed from an expression.

Constructors

PrimaryFieldAccess

Fields

infoPrimaryFieldAccess :: l

targetObject :: ExpNode l

targetField :: Ident

data SuperFieldAccess l

Accessing a field of the superclass.

Constructors

SuperFieldAccess

Fields

infoSuperFieldAccess :: l

superField :: Ident

data ClassFieldAccess l

Accessing a (static) field of a named class.

Constructors

ClassFieldAccess

Fields

infoClassFieldAccess :: l

targetClass :: Name

staticField :: Ident

data LambdaSingleParam l

Constructors

LambdaSingleParam

Fields

infoLambdaSingleParam :: l

lambdaParamName :: Ident

data LambdaFormalParams l

Constructors

LambdaFormalParams

Fields

infoLambdaFormalParams :: l

lambdaFormalParams :: [FormalParam l]

data LambdaInferredParams l

Constructors

LambdaInferredParams

Fields

infoLambdaInferredParams :: l

lambdaParamNames :: [Ident]

data MethodCall l

Invoking a specific named method.

Constructors

MethodCall

Fields

infoMethodCall :: l

methodCallName :: Name

methodCallArgs :: [Argument l]

data PrimaryMethodCall l

Invoking a method of a class computed from a primary expression, giving arguments for any generic type parameters.

Constructors

PrimaryMethodCall

Fields

infoPrimaryMethodCall :: l

methodCallTargetObj :: ExpNode l

mysteryRefTypes :: [RefType]

primaryMethodName :: Ident

primaryMethodCallArgs :: [Argument l]

data SuperMethodCall l

Invoking a method of the super class, giving arguments for any generic type parameters.

Constructors

SuperMethodCall

Fields

infoSuperMethodCall :: l

superMethodTypeArgs :: [RefType]

superMethodName :: Ident

superMethodArgs :: [Argument l]

data ClassMethodCall l

Invoking a method of the superclass of a named class, giving arguments for any generic type parameters.

Constructors

ClassMethodCall

Fields

infoClassMethodCall :: l

methodClassTarget :: Name

classMethodTypeArgs :: [RefType]

classMethodName :: Ident

classMethodArgs :: [Argument l]

data TypeMethodCall l

Invoking a method of a named type, giving arguments for any generic type parameters.

Constructors

TypeMethodCall

Fields

infoTypeMethodCall :: l

typeMethodClassTarget :: Name

typeMethodTypeArgs :: [RefType]

typeMethodName :: Ident

typeMethodArgs :: [Argument l]

data ArrayInit l

An array initializer may be specified in a declaration, or as part of an array creation expression, creating an array and providing some initial values

Constructors

ArrayInit

Fields

infoArrayInit :: l

arrayInits :: [VarInitNode l]

type Argument = ExpNode

Arguments to methods and constructors are expressions.