Language.Java.SyntaxClasses

Documentation

class HasType a where

Minimal complete definition

getType

Methods

getType :: a -> Type

class CollectTypes a where

Minimal complete definition

collectTypes

Methods

collectTypes :: a -> [Type]

class HasBody a l where

Provides functionality to access the body as a list of declarations of a class, enum and an interface.

Minimal complete definition

getBody

Methods

getBody :: a -> [DeclNode l]

getTypeFromPackage :: Package -> Type

class HasNode el no where

Minimal complete definition

toNode

Methods

toNode :: el l -> no l

Orphan instances

Eq l => Ord (MemberDeclNode l)

Methods

compare :: MemberDeclNode l -> MemberDeclNode l -> Ordering

(<) :: MemberDeclNode l -> MemberDeclNode l -> Bool

(<=) :: MemberDeclNode l -> MemberDeclNode l -> Bool

(>) :: MemberDeclNode l -> MemberDeclNode l -> Bool

(>=) :: MemberDeclNode l -> MemberDeclNode l -> Bool

max :: MemberDeclNode l -> MemberDeclNode l -> MemberDeclNode l

min :: MemberDeclNode l -> MemberDeclNode l -> MemberDeclNode l